opencode_3_tier_mem/mcp-servers/compchem-memory/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "compchem-memory"
version = "0.1.0"
description = "Three-tier memory MCP server for computational chemistry"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
